#3 Rearchitect Listener class
There is now a UartDwm1001 class that is the base of the Listener class.
This class provides UART-specific features such as a shell mode. The
shell mode is unsupported for other communication methods: SPI and I2C.
The UartDwm1001 base class provides core command sending functionality,
and the Listener class now essentially wraps the base with some more
intuitive functions.